fast-api/pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Vankad24 <vankad24@mail.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
